fix: rename navbar to Navbar, add proxy.ts (review fix)
